ASNIDE-161 Moved Data::Type child classes to dedicated directory and namespace
