+++ data/typeassignment.h
@@ -31,7 +31,7 @@
 #include "sourcelocation.h"
 #include "typereference.h"
 #include "node.h"
-#include "type.h"
+#include "types/type.h"
 
 namespace Asn1Acn {
 namespace Internal {
@@ -40,17 +40,17 @@ namespace Data {
 class TypeAssignment : public Node
 {
 public:
-    TypeAssignment(const QString &name, const SourceLocation &location, std::unique_ptr<Type> type);
+    TypeAssignment(const QString &name, const SourceLocation &location, std::unique_ptr<Types::Type> type);
     ~TypeAssignment() override;
 
     void accept(Visitor &visitor) const override;
 
     const QString &name() const { return m_name; }
-    const Type *type() const { return m_type.get(); }
+    const Types::Type *type() const { return m_type.get(); }
 
 private:
     QString m_name;
-    std::unique_ptr<Type> m_type;
+    std::unique_ptr<Types::Type> m_type;
 };
 
 } // namespace Data
